Added code for puzzle. Not yet done
There are now blocks on the wall and they sequence of being shot is being saved in a list.  Will compare the shot list to the generated list to see if they match. Also added canvas that display the blocks you have shot. It was difficult to see if you had hit it or not.